Rebase and respond to PR feedback.
