+++ src/Native/System.Security.Cryptography.Native/pal_crypto_config.h.in
@@ -2,3 +2,5 @@
 
 #cmakedefine01 HAVE_TLS_V1_1
 #cmakedefine01 HAVE_TLS_V1_2
+
+#cmakedefine01 HAVE_SSL_CIPHER_SPLIT_ALGORITHMS
