+++ src/System.Net.Security/src/System.Net.Security.csproj
@@ -269,9 +269,6 @@
     <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\Asn1SafeHandles.Unix.cs">
       <Link>Common\Microsoft\Win32\SafeHandles\Asn1SafeHandles.Unix.cs</Link>
     </Compile>
-    <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\SafeSslHandle.Unix.cs">
-      <Link>Common\Microsoft\Win32\SafeHandles\SafeSslHandle.Unix.cs</Link>
-    </Compile>
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
